feat: :truck: switch to new brand 'Brevo'
Changed all text referring to "Sendinblue" to "Brevo" including links, emails and apis endpoints. In the "setUserAgent" functions, the prefix "brevo_" was added as a parameter in the `.start_with?` function to support the new brand and maintain backward compatibility with the previous brand. In essence this commit does not break gem compatibility.